% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_diagonalCOV.R
\name{EM_diagonalCOV}
\alias{EM_diagonalCOV}
\title{Fit the GPCCA Model with Simplified Covariance Structure}
\usage{
EM_diagonalCOV(
  X.list,
  d = 2,
  tol = 1e-04,
  maxiter = 100,
  EarlyStop = TRUE,
  niter.ES = 5,
  W.init = "PCA",
  verbose = 1,
  plot_RMSE = TRUE
)
}
\arguments{
\item{X.list}{a list of numeric matrices. A multi-modal dataset of \eqn{R} 
modalities, where each matrix \eqn{\mathbf{X}^{(r)}} is a data modality, 
with \eqn{m_r} rows denoting features and \eqn{n} columns denoting samples 
(\eqn{1 \le r \le R}). All modalities must have the same sample size \eqn{n}, 
with the ordering of samples matched to each other. Artificial modality-wise 
missing values should be introduced by the user if non-matching samples are present.}

\item{d}{an integer. The size of target dimension \eqn{d}, i.e. the number of 
latent factors in low-dimensional subspace. (Default: \code{2})}

\item{tol}{a numeric scalar. Tolerance of the RMSE measuring the difference 
in the matrix of latent factors \eqn{\mathbf{Z}} between two consecutive 
iterations. This tolerance threshold \eqn{\tau} determines the stopping rule 
of the EM algorithm, and the RMSE is used to monitor the convergence.
(Default: \code{0.0001})}

\item{maxiter}{an integer. The maximum number of iterations allowed in the 
EM algorithm. (Default: \code{100})}

\item{EarlyStop}{a logical. Should early stop be used? If set to \code{TRUE}, 
it stops the EM algorithm early. With early stop turned on, only sub-optimal 
solution may be attained, but it helps avoid overfitting. (Default: \code{TRUE})}

\item{niter.ES}{an integer. The number of consecutive increase of RMSE allowed 
to decide for early stop. It affects nothing if \code{EarlyStop = FALSE}. 
(Default: \code{5})}

\item{W.init}{One of \code{"PCA"} and \code{"rand.std.norm"}, 
indicating which method is used for initializing the loading matrix 
\eqn{\mathbf{W}}. \code{"rand.std.norm"} initializes all elements of 
\eqn{\mathbf{W}} with i.i.d. standard normal random variates. \code{"PCA"} 
uses Randomized Principal Component Analysis (RPCA) if all data modalities 
are complete without missingness or Probabilistic Principal Component Analysis 
(PPCA) if any missing value is present in the multi-modal data \code{X.list}. 
(Default: \code{"PCA"})}

\item{verbose}{a binary. If set to \code{0}, it prevents the printing of 
runtime and RMSE in every iteration. (Default: \code{1})}

\item{plot_RMSE}{a logical. After the EM algorithm is completed, should the 
RMSEs in iterations be plotted? (Default: \code{TRUE})}
}
\value{
A list of 4 is returned, including:
\describe{
  \item{\code{Z}}{a numeric matrix; a matrix of size \eqn{d \times n} that 
    stores the fitted latent factors, i.e. the joint low-dimensional 
    embeddings of the original multi-modal data.}
  \item{\code{mu}}{a numeric vector; the estimated mean vector 
    \eqn{\boldsymbol{\mu}} of length \eqn{m}.}
  \item{\code{W}}{a numeric matrix; the estimated loading matrix 
    \eqn{\mathbf{W}} of size \eqn{m \times d}.}
  \item{\code{Psi}}{a sparse \code{Matrix}; the estimated error covariance matrix 
    \eqn{\mathbf{\Psi}} of size \eqn{m \times m}.}
}
}
\description{
Implementation of Generalized Probabilistic Canonical Correlation Analysis (GPCCA) 
  on a general multi-modal dataset that may contain missing values, if the 
  error covariance matrix \eqn{\mathbf{\Psi}} is assumed to be strict diagonal. 
  It is recommended for use when the total dimension of data (with all 
  modalities combined) is extremely high. In this case, ridge regularization 
  is not applied, without needing to specify the value of \eqn{\lambda}. 
  A sub-function invoked by the primary function \code{GPCCAmodel}. 
  See \code{\link{GPCCAmodel}} for more details of the GPCCA model.
}
\examples{
\dontrun{

## Generate a 3-modality continuous dataset with 20\% MCARs
DAT <- example_MultiModalData_sim(dataType = "continuous", missVal = TRUE, propMCAR = 0.2)

## Fit GPCCA model to the incomplete example multi-modal dataset
GPCCA.fit <- EM_diagonalCOV(X.list = DAT, d = 4)
                            
## Extract the fitted latent factors
LFs <- t(GPCCA.fit$Z)

## Data visualization
lbls <- example_MultiModalData_labels

par(pty = "s", xpd = TRUE, mar = c(5.1, 4.1, 4.1, 4.1))
plot(x = LFs[, 1], y = LFs[, 2], col = factor(lbls),
     type = "p", pch = 16, cex = 0.8, asp = 1,
     xlab = "Latent Factor 1", ylab = "Latent Factor 2",
     main = "Colored by ground truth labels")
legend("right", inset = c(-0.3, 0), title = "Group",
       legend = levels(factor(lbls)), col = 1:6,
       pch = 16, cex = 0.8)
}
}
\seealso{
\code{\link{GPCCAmodel}}, \code{\link{EM_completeDAT}}, \code{\link{EM_missingDAT}}
}
