% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_missingDAT.R
\name{EM_missingDAT}
\alias{EM_missingDAT}
\title{Fit the GPCCA Model on a General Multi-modal Dataset}
\usage{
EM_missingDAT(
  X.list,
  d = 2,
  lambda = 0.5,
  tol = 1e-04,
  maxiter = 100,
  EarlyStop = TRUE,
  niter.ES = 5,
  force_blkInv = FALSE,
  W.init = "PCA",
  verbose = 1,
  plot_RMSE = TRUE
)
}
\arguments{
\item{X.list}{a list of numeric matrices. A multi-modal dataset of \eqn{R} 
modalities, where each matrix \eqn{\mathbf{X}^{(r)}} is a data modality, 
with \eqn{m_r} rows denoting features and \eqn{n} columns denoting samples 
(\eqn{1 \le r \le R}). All modalities must have the same sample size \eqn{n}, 
with the ordering of samples matched to each other. Artificial modality-wise 
missing values should be introduced by the user if non-matching samples are present.}

\item{d}{an integer. The size of target dimension \eqn{d}, i.e. the number of 
latent factors in low-dimensional subspace. (Default: \code{2})}

\item{lambda}{a numeric scalar. The ridge regularization parameter \eqn{\lambda}, 
in the range of \eqn{(0, 1]}. A smaller \eqn{\lambda} corresponds to heavier 
penalty, which leads to a more sparse (diagonal-like) structure of the 
error covariance matrix \eqn{\mathbf{\Psi}}. In the extreme case, 
\eqn{\lambda = 1} means no ridge regularization is applied. 
It affects nothing if \code{diagCov = TRUE}. (Default: \code{0.5})}

\item{tol}{a numeric scalar. Tolerance of the RMSE measuring the difference 
in the matrix of latent factors \eqn{\mathbf{Z}} between two consecutive 
iterations. This tolerance threshold \eqn{\tau} determines the stopping rule 
of the EM algorithm, and the RMSE is used to monitor the convergence.
(Default: \code{0.0001})}

\item{maxiter}{an integer. The maximum number of iterations allowed in the 
EM algorithm. (Default: \code{100})}

\item{EarlyStop}{a logical. Should early stop be used? If set to \code{TRUE}, 
it stops the EM algorithm early. With early stop turned on, only sub-optimal 
solution may be attained, but it helps avoid overfitting. (Default: \code{TRUE})}

\item{niter.ES}{an integer. The number of consecutive increase of RMSE allowed 
to decide for early stop. It affects nothing if \code{EarlyStop = FALSE}. 
(Default: \code{5})}

\item{force_blkInv}{a logical. Should blockwise inversion of the error 
covariance matrix \eqn{\mathbf{\Psi}} be activated? A technique to boost 
computation. It is automatically turned on if either the dimension of any 
data modality \eqn{m_r} is larger than \eqn{500} or the total dimension 
\eqn{m = \sum_{r=1}^{R}m_r} is larger than \eqn{1000}. 
It affects nothing if \code{diagCov = TRUE}. (Default: \code{FALSE})}

\item{W.init}{One of \code{"PCA"} and \code{"rand.std.norm"}, 
indicating which method is used for initializing the loading matrix 
\eqn{\mathbf{W}}. \code{"rand.std.norm"} initializes all elements of 
\eqn{\mathbf{W}} with i.i.d. standard normal random variates. \code{"PCA"} 
uses Randomized Principal Component Analysis (RPCA) if all data modalities 
are complete without missingness or Probabilistic Principal Component Analysis 
(PPCA) if any missing value is present in the multi-modal data \code{X.list}. 
(Default: \code{"PCA"})}

\item{verbose}{a binary. If set to \code{0}, it prevents the printing of 
runtime and RMSE in every iteration. (Default: \code{1})}

\item{plot_RMSE}{a logical. After the EM algorithm is completed, should the 
RMSEs in iterations be plotted? (Default: \code{TRUE})}
}
\value{
A list of 4 is returned, including:
\describe{
  \item{\code{Z}}{a numeric matrix; a matrix of size \eqn{d \times n} that 
    stores the fitted latent factors, i.e. the joint low-dimensional 
    embeddings of the original multi-modal data.}
  \item{\code{mu}}{a numeric vector; the estimated mean vector 
    \eqn{\boldsymbol{\mu}} of length \eqn{m}.}
  \item{\code{W}}{a numeric matrix; the estimated loading matrix 
    \eqn{\mathbf{W}} of size \eqn{m \times d}.}
  \item{\code{Psi}}{a sparse \code{Matrix}; the estimated error covariance matrix 
    \eqn{\mathbf{\Psi}} of size \eqn{m \times m}.}
}
}
\description{
Implementation of Generalized Probabilistic Canonical Correlation Analysis (GPCCA) 
  on a general multi-modal dataset that may contain missing values. 
  A sub-function invoked by the primary function \code{GPCCAmodel}. 
  See \code{\link{GPCCAmodel}} for more details of the GPCCA model.
}
\examples{
\dontrun{

## Generate a 3-modality continuous dataset with 20\% MCARs
DAT <- example_MultiModalData_sim(dataType = "continuous", missVal = TRUE, propMCAR = 0.2)

## Fit GPCCA model to the incomplete example multi-modal dataset
GPCCA.fit <- EM_missingDAT(X.list = DAT, d = 4, lambda = 0.5)
                           
## Extract the fitted latent factors
LFs <- t(GPCCA.fit$Z)

## Data visualization
lbls <- example_MultiModalData_labels

par(pty = "s", xpd = TRUE, mar = c(5.1, 4.1, 4.1, 4.1))
plot(x = LFs[, 1], y = LFs[, 2], col = factor(lbls),
     type = "p", pch = 16, cex = 0.8, asp = 1,
     xlab = "Latent Factor 1", ylab = "Latent Factor 2",
     main = "Colored by ground truth labels")
legend("right", inset = c(-0.3, 0), title = "Group",
       legend = levels(factor(lbls)), col = 1:6,
       pch = 16, cex = 0.8)
}
}
\seealso{
\code{\link{GPCCAmodel}}, \code{\link{EM_completeDAT}}, \code{\link{EM_diagonalCOV}}
}
