% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_missingDAT_Pois.R
\name{EM_missingDAT_Pois}
\alias{EM_missingDAT_Pois}
\title{Fit the GPCCA-Pois Model on a General Multi-modal Dataset of Counts}
\usage{
EM_missingDAT_Pois(
  Y.list,
  d = 2,
  lambda = 0.1,
  tol = 1e-04,
  maxiter = 50,
  EarlyStop = TRUE,
  niter.ES = 5,
  W.init = "PCA",
  verbose = 1,
  plot_RMSE = TRUE
)
}
\arguments{
\item{Y.list}{a list of numeric matrices. A multi-modal dataset of \eqn{R} 
modalities, where each matrix \eqn{\mathbf{Y}^{(r)}} is a data modality, 
with \eqn{m_r} rows denoting features and \eqn{n} columns denoting samples 
(\eqn{1 \le r \le R}). All modalities must have the same sample size \eqn{n}, 
with the ordering of samples matched to each other. Artificial modality-wise 
missing values should be introduced by the user if non-matching samples are present.}

\item{d}{an integer. The size of target dimension \eqn{d}, i.e. the number of 
latent factors in low-dimensional subspace. (Default: \code{2})}

\item{lambda}{a numeric scalar. The ridge regularization parameter \eqn{\lambda}, 
in the range of \eqn{[0, \infty)}. A larger \eqn{\lambda} corresponds to heavier 
penalty, which leads to a more unique solution of the loading matrix \eqn{\mathbf{W}} 
and correspondingly more unique joint low-dimensional embeddings. 
\eqn{\lambda = 0} means no ridge regularization is applied. (Default: \code{0.1})}

\item{tol}{a numeric scalar. Tolerance of the RMSE measuring the difference 
in the matrix of latent factors \eqn{\mathbf{Z}} between two consecutive 
iterations. This tolerance threshold \eqn{\tau} determines the stopping rule 
of the EM algorithm, and the RMSE is used to monitor the convergence.
(Default: \code{0.0001})}

\item{maxiter}{an integer. The maximum number of iterations allowed in the 
EM algorithm. (Default: \code{50})}

\item{EarlyStop}{a logical. Should early stop be used? If set to \code{TRUE}, 
it stops the EM algorithm early. With early stop turned on, only sub-optimal 
solution may be attained, but it helps avoid overfitting. (Default: \code{TRUE})}

\item{niter.ES}{an integer. The number of consecutive increase of RMSE allowed 
to decide for early stop. It affects nothing if \code{EarlyStop = FALSE}. 
(Default: \code{5})}

\item{W.init}{One of \code{"PCA"} and \code{"rand.std.norm"}, 
indicating which method is used for initializing the loading matrix 
\eqn{\mathbf{W}}. \code{"rand.std.norm"} initializes all elements of 
\eqn{\mathbf{W}} with i.i.d. standard normal random variates. \code{"PCA"} 
uses Randomized Principal Component Analysis (RPCA) if all data modalities 
are complete without missingness or Probabilistic Principal Component Analysis 
(PPCA) if any missing value is present in the multi-modal data \code{Y.list}. 
(Default: \code{"PCA"})}

\item{verbose}{a binary. If set to \code{0}, it prevents the printing of 
runtime and RMSE in every iteration. (Default: \code{1})}

\item{plot_RMSE}{a logical. After the EM algorithm is completed, should the 
RMSEs in iterations be plotted? (Default: \code{TRUE})}
}
\value{
A list of 3 is returned, including:
\describe{
  \item{\code{Z}}{a numeric matrix; a matrix of size \eqn{d \times n} that 
    stores the fitted latent factors, i.e. the joint low-dimensional 
    embeddings of the original multi-modal data.}
  \item{\code{mu}}{a numeric vector; the estimated mean vector 
    \eqn{\boldsymbol{\mu}} of length \eqn{m}.}
  \item{\code{W}}{a numeric matrix; the estimated loading matrix 
    \eqn{\mathbf{W}} of size \eqn{m \times d}.}
}
}
\description{
Implementation of GPCCA-Pois on a general multi-modal dataset of counts 
  that may contain missing values. 
  A sub-function invoked by the primary function \code{GPCCAmodel2}. 
  See \code{\link{GPCCAmodel2}} for more details of the extension models from GPCCA.
}
\examples{
\dontrun{

## Generate a 3-modality count dataset with 20\% MCARs
DAT <- example_MultiModalData_sim(dataType = "count", missVal = TRUE, propMCAR = 0.2)

## Fit GPCCA-Pois model to the incomplete example multi-modal dataset
GPCCA.fit <- EM_missingDAT_Pois(Y.list = DAT, d = 4, lambda = 0.1)

## Extract the fitted latent factors
LFs <- t(GPCCA.fit$Z)

## Data visualization
lbls <- example_MultiModalData_labels

par(pty = "s", xpd = TRUE, mar = c(5.1, 4.1, 4.1, 4.1))
plot(x = LFs[, 1], y = LFs[, 2], col = factor(lbls),
     type = "p", pch = 16, cex = 0.8, asp = 1,
     xlab = "Latent Factor 1", ylab = "Latent Factor 2",
     main = "Colored by ground truth labels")
legend("right", inset = c(-0.3, 0), title = "Group",
       legend = levels(factor(lbls)), col = 1:6,
       pch = 16, cex = 0.8)
}
}
\seealso{
\code{\link{GPCCAmodel2}}, \code{\link{EM_completeDAT_Pois}}
}
