% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_MultiModalData_labels}
\alias{example_MultiModalData_labels}
\title{Ground Truth Labels of Example Multi-modal Dataset}
\format{
A character vector that consists of 50 \code{"group 1"}, 
50 \code{"group 2"}, ... , 50 \code{"group 6"} in order.
}
\usage{
example_MultiModalData_labels
}
\description{
Ground truth sample labels of the synthetic multi-modal dataset given by 
\code{example_MultiModalData_list0} (complete) or 
\code{example_MultiModalData_list1} (incomplete).
}
\seealso{
\code{\link{example_MultiModalData_list0}}, 
         \code{\link{example_MultiModalData_list1}}
}
\keyword{datasets}
