% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_MultiModalData_list0}
\alias{example_MultiModalData_list0}
\title{Example Multi-modal Dataset (complete)}
\format{
A list that consists of 3 numeric matrices:
\describe{
  \item{M1}{Modality 1: a matrix of size \eqn{m_1 \times n}, 
    i.e. 30 features (rows) by 300 samples (columns)}
  \item{M2}{Modality 2: a matrix of size \eqn{m_2 \times n}, 
    i.e. 60 features (rows) by 300 samples (columns)}
  \item{M3}{Modality 3: a matrix of size \eqn{m_3 \times n}, 
    i.e. 90 features (rows) by 300 samples (columns)}
}
}
\usage{
example_MultiModalData_list0
}
\description{
A synthetic multi-modal dataset with three modalities. The number of features 
  in each modality is \eqn{m_1 = 30}, \eqn{m_2 = 60} and \eqn{m_3 = 90}, 
  respectively. There are six groups of samples where each group has a equal 
  sample size of \eqn{50}. In total, there are \eqn{n = 300} samples, 
  with ground truth labels provided by \code{example_MultiModalData_labels}. 
  The dataset is complete with no missing value.
}
\seealso{
\code{\link{example_MultiModalData_list1}}, 
         \code{\link{example_MultiModalData_labels}}
}
\keyword{datasets}
