% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_MultiModalData_sim.R
\name{example_MultiModalData_sim}
\alias{example_MultiModalData_sim}
\title{Generate Example Multi-modal Dataset}
\usage{
example_MultiModalData_sim(
  dataType = "continuous",
  missVal = TRUE,
  propMCAR = 0.2
)
}
\arguments{
\item{dataType}{One of \code{"continuous"}, \code{"count"} and \code{"binary"}, 
specifying the data type of matrices to be generated. (Default: \code{"continuous"})}

\item{missVal}{a logical. 
Should artificial missing values (\code{NA}s) be introduced? (Default: \code{TRUE})}

\item{propMCAR}{a numeric scalar. 
The overall missing rate due to MCAR, between \code{0} and \code{0.5}. 
It affects nothing if \code{missVal = FALSE}. (Default: \code{0.2})}
}
\value{
A list that consists of 3 numeric matrices is returned:
\describe{
  \item{M1}{Modality 1: a matrix of size \eqn{m_1 \times n}, 
    i.e. 30 features (rows) by 300 samples (columns).}
  \item{M2}{Modality 2: a matrix of size \eqn{m_2 \times n}, 
    i.e. 60 features (rows) by 300 samples (columns).}
  \item{M3}{Modality 3: a matrix of size \eqn{m_3 \times n}, 
    i.e. 90 features (rows) by 300 samples (columns).}
}
}
\description{
Simulation of a synthetic multi-modal dataset with three modalities. 
  The number of features in each modality is \eqn{m_1 = 30}, \eqn{m_2 = 60} and \eqn{m_3 = 90}, 
  respectively. There are six groups of samples where each group has a equal 
  sample size of \eqn{50}. In total, there are \eqn{n = 300} samples, 
  with ground truth labels provided by \code{example_MultiModalData_labels}. 
  Artificial missing values are introduced with MCAR mechanism. 
  The simulated datasets should vary under different random seeds.
}
\examples{
\dontrun{

## I: Generate a 3-modality continuous dataset with no missing value
DAT <- example_MultiModalData_sim(dataType = "continuous", missVal = FALSE)

## II: Generate a 3-modality count dataset with 10\% MCARs
DAT <- example_MultiModalData_sim(dataType = "count", missVal = TRUE, propMCAR = 0.1)

## III: Generate a 3-modality binary dataset with 50\% MCARs
DAT <- example_MultiModalData_sim(dataType = "binary", missVal = TRUE, propMCAR = 0.5)
}
}
\seealso{
\code{\link{example_MultiModalData_labels}}
}
